1125_final_pjt/final-pjt-back/server/requirements.txt
asgiref==3.4.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==3.2.6
django-cors-headers==3.13.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
idna==3.4
joblib==1.2.0
JPype1==1.4.1
konlpy==0.6.0
krwordrank==1.0.3
lxml==4.9.1
numpy==1.23.5
packaging==21.3
PyJWT==2.6.0
pyparsing==3.0.9
pytz==2021.1
requests==2.28.1
scikit-learn==1.1.3
scipy==1.9.3
sqlparse==0.4.1
threadpoolctl==3.1.0
urllib3==1.26.12
